chore!: refactor MapeoManager to accept Fastify server instance (#440)
