+++ src/index.js
@@ -1,1 +1,2 @@
+export { FastifyController } from './fastify-controller.js'
 export { MapeoManager } from './mapeo-manager.js'
