+++ src/mapeo-manager.js
@@ -11,6 +11,7 @@ import { TypedEmitter } from 'tiny-typed-emitter'
 import { IndexWriter } from './index-writer/index.js'
 import {
   MapeoProject,
+  kBlobStore,
   kProjectLeave,
   kSetOwnDeviceInfo,
 } from './mapeo-project.js'
@@ -30,9 +31,11 @@ import {
   projectKeyToPublicId,
 } from './utils.js'
 import { RandomAccessFilePool } from './core-manager/random-access-file-pool.js'
+import BlobServerPlugin from './fastify-plugins/blobs.js'
+import IconServerPlugin from './fastify-plugins/icons.js'
+import { getFastifyServerAddress } from './fastify-plugins/utils.js'
 import { LocalPeers } from './local-peers.js'
 import { InviteApi } from './invite-api.js'
-import { MediaServer } from './media-server.js'
 import { LocalDiscovery } from './discovery/local-discovery.js'
 import { Capabilities } from './capabilities.js'
 import NoiseSecretStream from '@hyperswarm/secret-stream'
@@ -49,6 +52,10 @@ const CLIENT_SQLITE_FILE_NAME = 'client.db'
 // limit of 1024 per process, so choosing 768 to leave 256 descriptors free for
 // other things e.g. SQLite and other parts of the app.
 const MAX_FILE_DESCRIPTORS = 768
+
+// Prefix names for routes registered with http server
+const BLOBS_PREFIX = 'blobs'
+const ICONS_PREFIX = 'icons'
 
 export const kRPC = Symbol('rpc')
 export const kManagerReplicate = Symbol('replicate manager')
@@ -80,7 +87,7 @@ export class MapeoManager extends TypedEmitter {
   #deviceId
   #localPeers
   #invite
-  #mediaServer
+  #fastify
   #localDiscovery
   #loggerBase
   #l
@@ -92,7 +99,7 @@ export class MapeoManager extends TypedEmitter {
    * @param {string} opts.projectMigrationsFolder path for drizzle migrations folder for project database
    * @param {string} opts.clientMigrationsFolder path for drizzle migrations folder for client database
    * @param {string | import('./types.js').CoreStorage} opts.coreStorage Folder for hypercore storage or a function that returns a RandomAccessStorage instance
-   * @param {{ port?: number, logger: import('fastify').FastifyServerOptions['logger'] }} [opts.mediaServerOpts]
+   * @param {import('fastify').FastifyInstance} opts.fastify Fastify server instance
    */
   constructor({
     rootKey,
@@ -100,7 +107,7 @@ export class MapeoManager extends TypedEmitter {
     projectMigrationsFolder,
     clientMigrationsFolder,
     coreStorage,
-    mediaServerOpts,
+    fastify,
   }) {
     super()
     this.#keyManager = new KeyManager(rootKey)
@@ -160,8 +167,16 @@ export class MapeoManager extends TypedEmitter {
       this.#coreStorage = coreStorage
     }
 
-    this.#mediaServer = new MediaServer({
-      logger: mediaServerOpts?.logger,
+    this.#fastify = fastify
+    this.#fastify.register(BlobServerPlugin, {
+      prefix: BLOBS_PREFIX,
+      getBlobStore: async (projectPublicId) => {
+        const project = await this.getProject(projectPublicId)
+        return project[kBlobStore]
+      },
+    })
+    this.#fastify.register(IconServerPlugin, {
+      prefix: ICONS_PREFIX,
       getProject: this.getProject.bind(this),
     })
 
@@ -197,6 +212,35 @@ export class MapeoManager extends TypedEmitter {
       keyPair: this.#keyManager.getIdentityKeypair(),
     })
     return this.#replicate(noiseStream)
+  }
+
+  /**
+   * @param {'blobs' | 'icons' | 'maps'} mediaType
+   * @returns {Promise<string>}
+   */
+  async #getMediaBaseUrl(mediaType) {
+    /** @type {string | null} */
+    let prefix = null
+
+    switch (mediaType) {
+      case 'blobs': {
+        prefix = BLOBS_PREFIX
+        break
+      }
+      case 'icons': {
+        prefix = ICONS_PREFIX
+        break
+      }
+      default: {
+        throw new Error(`Unsupported media type ${mediaType}`)
+      }
+    }
+
+    const base = await getFastifyServerAddress(this.#fastify.server, {
+      timeout: 5000,
+    })
+
+    return base + '/' + prefix
   }
 
   /**
@@ -403,9 +447,7 @@ export class MapeoManager extends TypedEmitter {
       sharedIndexWriter: this.#projectSettingsIndexWriter,
       localPeers: this.#localPeers,
       logger: this.#loggerBase,
-      getMediaBaseUrl: this.#mediaServer.getMediaAddress.bind(
-        this.#mediaServer
-      ),
+      getMediaBaseUrl: this.#getMediaBaseUrl.bind(this),
     })
   }
 
@@ -652,17 +694,6 @@ export class MapeoManager extends TypedEmitter {
    */
   get invite() {
     return this.#invite
-  }
-
-  /**
-   * @param {import('./media-server.js').StartOpts} [opts]
-   */
-  async startMediaServer(opts) {
-    await this.#mediaServer.start(opts)
-  }
-
-  async stopMediaServer() {
-    await this.#mediaServer.stop()
   }
 
   async startLocalPeerDiscovery() {
